refactor request reset page to separate file
