+++ www/src/components/users/RequestPasswordReset.tsx
@@ -1,0 +1,113 @@
+import { useState } from 'react'
+import { Link, useLocation } from 'react-router-dom'
+import { Form } from 'grommet'
+import {
+  A,
+  Button,
+  Div,
+  Flex,
+  H1,
+  P,
+  Span,
+} from 'honorable'
+import { SuccessIcon } from '@pluralsh/design-system'
+
+import { ResetTokenType, useCreateResetTokenMutation } from '../../generated/graphql'
+import { isMinViableEmail } from '../../utils/string'
+import { GqlError } from '../utils/Alert'
+
+import { LabelledInput, LoginPortal } from './MagicLogin'
+
+export function RequestPasswordReset() {
+  const location = useLocation()
+  const [attributes, setAttributes] = useState({
+    email: location?.state?.email || '',
+    type: ResetTokenType.Password,
+  })
+  const [mutation, { loading, data, error }] = useCreateResetTokenMutation({
+    variables: { attributes },
+  })
+
+  const resetSuccess = data && data.createResetToken
+  const invalidEmail = error?.message === 'not_found'
+  const gqlError = !invalidEmail ? error : undefined
+
+  return (
+    <LoginPortal>
+      {resetSuccess ? (
+        <RequestResetSuccess />
+      ) : (
+        <>
+          <Div marginBottom="xlarge">
+            <H1
+              title1
+              textAlign="center"
+            >
+              Reset password
+            </H1>
+          </Div>
+          {gqlError && (
+            <Div marginBottom="medium">
+              <GqlError
+                header="Failed!"
+                error={error}
+              />
+            </Div>
+          )}
+          <Form onSubmit={() => mutation()}>
+            <LabelledInput
+              width="100%"
+              value={attributes.email}
+              label="Email"
+              name="email"
+              placeholder="your email"
+              onChange={email => setAttributes({ ...attributes, email })}
+              hint={
+                invalidEmail ? (
+                  <Span color="text-danger-light">Invalid email address</Span>
+                ) : undefined
+              }
+              inputProps={{ error: invalidEmail }}
+            />
+            <Button
+              type="submit"
+              width="100%"
+              loading={loading}
+              disabled={!isMinViableEmail(attributes.email)}
+            >
+              Reset Password
+            </Button>
+          </Form>
+        </>
+      )}
+    </LoginPortal>
+  )
+}
+
+function RequestResetSuccess() {
+  return (
+    <Flex
+      flexDirection="column"
+      alignItems="center"
+      justifyContent="center"
+      textAlign="center"
+    >
+      <SuccessIcon
+        color="icon-success"
+        size={32}
+        marginBottom="large"
+      />
+      <Div marginBottom="xlarge">
+        <H1 title1>Password reset email sent</H1>
+        <P>Check your email to continue the process.</P>
+      </Div>
+      <A
+        inline
+        as={Link}
+        to="/login"
+      >
+        Back to login
+      </A>
+    </Flex>
+  )
+}
